Organized app.go into multiple files
