--- playwright.config.js
@@ -1,13 +1,0 @@
-import { defineConfig, devices } from '@playwright/test';
-
-export default defineConfig({
-	testDir: './tests',
-	use: {
-		baseURL: 'http://localhost:8080',
-	},
-	projects: [
-		{ name: 'chromium', use: { ...devices['Desktop Chrome'] } },
-		{ name: 'firefox', use: { ...devices['Desktop Firefox'] } },
-		{ name: 'webkit', use: { ...devices['Desktop Safari'] } },
-	],
-});
