+++ src/backend/models.go
@@ -1,0 +1,38 @@
+package main
+
+import (
+	"time"
+)
+
+type User struct {
+	ID       int    `json:"id"`
+	Username string `json:"username"`
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
+type PageData struct {
+	User         *User
+	Error        string
+	Title        string
+	Template     string
+	UserLoggedIn bool
+}
+
+type Page struct {
+	Title       string    `json:"title"`
+	URL         string    `json:"url"`
+	Language    string    `json:"language"`
+	LastUpdated time.Time `json:"last_updated"`
+	Content     string    `json:"content"`
+}
+
+type WeatherResponse struct {
+	Name string `json:"name"`
+	Main struct {
+		Temp float64 `json:"temp"`
+	} `json:"main"`
+	Weather []struct {
+		Description string `json:"description"`
+	} `json:"weather"`
+}
