+++ src/backend/search.log
@@ -4,3 +4,4 @@ SEARCH: 2025/04/27 11:07:02 query="database" from=[::1]:50021
 SEARCH: 2025/04/27 11:10:18 query="tlou" from=[::1]:50021
 SEARCH: 2025/04/27 11:10:27 query="devops" from=[::1]:50021
 SEARCH: 2025/04/28 20:33:34 query="python" from=172.19.0.1:48652
+SEARCH: 2025/05/01 17:34:54 query="python" from=[::1]:60352
